[3D] Clean up DBC files
