+++ WDE.MapRenderer/Managers/DbcManager.cs
@@ -1,4 +1,5 @@
 using WDE.Common.DBC;
+using WDE.MpqReader.DBC;
 using WDE.MpqReader.Structures;
 
 namespace WDE.MapRenderer.Managers
