+++ WDE.MapRenderer/Managers/LightingManager.cs
@@ -4,6 +4,7 @@ using TheEngine.Handles;
 using TheEngine.Interfaces;
 using TheMaths;
 using WDE.MapRenderer.StaticData;
+using WDE.MpqReader.DBC;
 using WDE.MpqReader.Structures;
 
 namespace WDE.MapRenderer.Managers
