+++ WDE.MpqReader/DBC/AreaTriggerShape.cs
@@ -1,0 +1,7 @@
+namespace WDE.MpqReader.DBC;
+
+public enum AreaTriggerShape
+{
+    Box,
+    Sphere
+}
