+++ WDE.MpqReader/DBC/EmoteStore.cs
@@ -1,0 +1,15 @@
+using WDE.Common.DBC;
+
+namespace WDE.MpqReader.DBC;
+
+public class EmoteStore : BaseDbcStore<uint, Emote>
+{
+    public EmoteStore(IEnumerable<IDbcIterator> rows)
+    {
+        foreach (var row in rows)
+        {
+            var o = new Emote(row);
+            store[o.Id] = o;
+        }
+    }
+}
