+++ WDE.MpqReader/DBC/EmoteType.cs
@@ -1,0 +1,8 @@
+namespace WDE.MpqReader.DBC;
+
+public enum EmoteType
+{
+    OneShotEmote,
+    ChangeStandState,
+    SetEmoteState
+}
