+++ WDE.MpqReader/DBC/GameObjectDisplayInfo.cs
@@ -1,7 +1,6 @@
-﻿using System.Collections;
-using WDE.Common.DBC;
+﻿using WDE.Common.DBC;
 
-namespace WDE.MpqReader.Structures
+namespace WDE.MpqReader.DBC
 {
     public class GameObjectDisplayInfo
     {
@@ -47,23 +46,4 @@ namespace WDE.MpqReader.Structures
 
         public static GameObjectDisplayInfo Empty => new GameObjectDisplayInfo();
     }
-
-    public class GameObjectDisplayInfoStore : IEnumerable<GameObjectDisplayInfo>
-    {
-        private Dictionary<int, GameObjectDisplayInfo> store = new();
-        public GameObjectDisplayInfoStore(IEnumerable<IDbcIterator> rows)
-        {
-            foreach (var row in rows)
-            {
-                var o = new GameObjectDisplayInfo(row);
-                store[o.Id] = o;
-            }
-        }
-
-        public bool Contains(int id) => store.ContainsKey(id);
-        public GameObjectDisplayInfo this[int id] => store[id];
-        public IEnumerator<GameObjectDisplayInfo> GetEnumerator() => store.Values.GetEnumerator();
-        IEnumerator IEnumerable.GetEnumerator() => store.Values.GetEnumerator();
-    }
-
 }
