+++ WDE.MpqReader/DBC/LightFloatParamType.cs
@@ -1,0 +1,11 @@
+namespace WDE.MpqReader.DBC;
+
+public enum LightFloatParamType
+{
+    FogDistance = 0,
+    FogMultiplier = 1,
+    SunGodRays = 2,
+    CloudDensity = 3,
+    Unk1 = 4,
+    Unk2 = 5
+}
