+++ WDE.MpqReader/DBC/LightIntParamType.cs
@@ -1,0 +1,23 @@
+namespace WDE.MpqReader.DBC;
+
+public enum LightIntParamType
+{
+    GeneralLightning = 0,
+    AmbientLight = 1,
+    SkyTopMost = 2,
+    SkyMiddle = 3,
+    SkyToHorizon = 4,
+    SkyJustAboveHorizon = 5,
+    SkyHorizon = 6,
+    FogInTheBackground = 7,
+    Unknown1 = 8,
+    SunColor = 9,
+    Clouds1 = 10,
+    Clouds2 = 11,
+    Clouds3 = 12,
+    Unknown2 = 13,
+    OceanShallow = 14,
+    OceanDeep = 15,
+    RiverShallow = 16,
+    RiverDeep = 17
+}
