+++ WDE.MpqReader/DBC/MapStore.cs
@@ -1,31 +1,8 @@
 using System.Collections;
 using WDE.Common.DBC;
 
-namespace WDE.MpqReader.Structures
+namespace WDE.MpqReader.DBC
 {
-    public class Map
-    {
-        public readonly int Id;
-        public readonly string Directory;
-        public readonly string Name;
-
-        public Map(IDbcIterator dbcIterator)
-        {
-            Id = dbcIterator.GetInt(0);
-            Directory = dbcIterator.GetString(1);
-            Name = dbcIterator.GetString(5);
-        }
-
-        private Map()
-        {
-            Id = -1;
-            Directory = "";
-            Name = "(null)";
-        }
-
-        public static Map Empty => new Map();
-    }
-
     public class MapStore : IEnumerable<Map>
     {
         private Dictionary<int, Map> store = new();
